pyproject.toml
[project]
name = "code-roaster"
version = "0.1.0"
description = ""
authors = [
    {name = "Mohamed Amr",email = "mohameda.elfarouk@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi (>=0.128.0,<0.129.0)",
    "uvicorn[standard] (>=0.40.0,<0.41.0)",
    "pydantic-settings (>=2.12.0,<3.0.0)",
    "google-generativeai (>=0.8.6,<0.9.0)"
]

[tool.poetry]
packages = [{include = "code_roaster", from = "src"}]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
